+++ m8tes/_resources/users.py
@@ -78,6 +78,7 @@ class Users:
         return SyncPage(
             data=[EndUserUsage.from_dict(d) for d in body["data"]],
             has_more=body["has_more"],
+            next_starting_after=body.get("next_starting_after"),
             _fetch_next=_fetch_next,
         )
 
@@ -97,6 +98,7 @@ class Users:
         return SyncPage(
             data=[EndUser.from_dict(d) for d in body["data"]],
             has_more=body["has_more"],
+            next_starting_after=body.get("next_starting_after"),
             _fetch_next=_fetch_next,
         )