+++ m8tes/_resources/webhooks.py
@@ -91,6 +91,7 @@ class Webhooks:
         return SyncPage(
             data=[Webhook.from_dict(d) for d in body["data"]],
             has_more=body["has_more"],
+            next_starting_after=body.get("next_starting_after"),
             _fetch_next=_fetch_next,
         )
 
@@ -134,6 +135,7 @@ class Webhooks:
         return SyncPage(
             data=[WebhookDelivery.from_dict(d) for d in body["data"]],
             has_more=body["has_more"],
+            next_starting_after=body.get("next_starting_after"),
             _fetch_next=_fetch_next,
         )